migrate final sample app makeStyles
